Add support to attach kprobe to offset
Example: `bpftrace -e 'kprobe:do_sys_open+9 { ... }'`.
This is based on the support of attaching uprobe to offset (#803).  The
offset is checked whether it is an instruction boundary using vmlinux.
The environment variable `BPFTRACE_VMLINUX` can be used to specify
vmlinux file.  ALLOW_UNSAFE_UPROBE option is renamed to
ALLOW_UNSASFE_PROBE so that it supports both uprobe and kprobe.