Error if trying to store context in a map
The verifier will reject instructions trying to store the context in a
map. Print an error message about it in the semantic analyzer. e.g.,

```
% bpftrace -e 'kprobe:f { @ = ctx; }'
stdin:1:14-15: ERROR: Storing context in the map is not supported
kprobe:f { @ = ctx; }
             ~
```

Limitation: This does not work if manually casting to an integer. e.g.,

```
% bpftrace -e 'kprobe:f { @ = (uint64)ctx; }'
Attaching 1 probe...
Error loading program: kprobe:f (try -v)
```

This is just the same as the previous behavior. One of the solution to
this is prohibiting casting Type::ctx to Type::integer. For now it is
allowed for some workaround.

------

Closes #978