Skip is_traceable_func() when fuzzing
I realized that reading debug/tracing/available_filter_functions takes
time, and that slow down fuzzing severely. Especially, BTF's constructor
always call the function no matter if it does not use the function. So this
patches changes it, and now BTF's constructor does not call the
function, and always report tracee function is available. It should be
OK since during fuzzing we don't load the program in the kernel and
trace. Now average AFL's exec speed becomes from 1 to up to 20 in my machine.